Move some patches to PatchEngine
