+++ MWSE/MWSE.vcxproj
@@ -38,6 +38,7 @@
     <ClInclude Include="PatchActors.h" />
     <ClInclude Include="PatchAudio.h" />
     <ClInclude Include="PatchDialogue.h" />
+    <ClInclude Include="PatchEngine.h" />
     <ClInclude Include="PatchInput.h" />
     <ClInclude Include="PatchIO.h" />
     <ClInclude Include="PatchMagic.h" />
@@ -702,6 +703,7 @@
     <ClCompile Include="PatchActors.cpp" />
     <ClCompile Include="PatchAudio.cpp" />
     <ClCompile Include="PatchDialogue.cpp" />
+    <ClCompile Include="PatchEngine.cpp" />
     <ClCompile Include="PatchInput.cpp" />
     <ClCompile Include="PatchIO.cpp" />
     <ClCompile Include="PatchMagic.cpp" />
