+++ MWSE/PatchEngine.h
@@ -1,0 +1,7 @@
+#pragma once
+
+namespace mwse::patch::engine {
+	void install();
+	void installPostLua();
+	void installPostInitialization();
+}
